Create recieve_info_python
here, the python file will recieve an array given by the object sys. The len of the array are going to be the number of information separeted with " "(spaces) given in the string by the csharp file. For example, if the string is "pyhon csharp connection", sys[0] == 'python', sys[1] == 'csharp' and sys[2] == 'connection'